Scale machineset library
A common use case across midstream repos is to scale up the worker
nodes.

The way it has been done so far is to always set whatever is the
first machineset to a specified number of workers. Obviously, this
isn't ideal since the sum of worker nodes across the (usual)
multiple machineset might already meet the expected number of
replicas.

Signed-off-by: Pierangelo Di Pilato <pierdipi@redhat.com>