testselect for selecting tests based on clonerefs and testsuites yaml (#27)
* Read and print out clonerefs.json and testsuites.yaml

* Generate a list of changed files from two SHAs

* Select working, output sent to file

*Still having "upstream" flag available.

* Get rid of upstream flag, sort keys

* Remove redundant comments and json support for testsuites.yaml

* Handle case when Pulls are not available

* Use Git clone->checkout->fetch->merge->diff sequence

The previous approach was giving wrong results when the branch from pull
request was behind the base branch (not rebased).

This new approach performs merge before doing the diff.

* Tests is just a slice of strings

* Cleanup

* Test suites without run_if_changed run always

* Fix formatting

* Fix formatting #2

* Be explicit about which tests should run even if All is added

This allows for running tests only if they were explicitly mentioned as
opposed to running them in a general case when some "unknown" path
changes.

* Add comment about the meaning of Tests field