Allow skipping must-gather on success
For periodic jobs, we'd be running must-gather always - on both test
failure/success.
For pre-submit jobs, the must-gather would be run only on failure.

Background: Gathering artifacts on success is useful only rarely, e.g.
for comparing failed and successful runs. But one can look at the
periodic job to get an idea what the successful run looks like. There's
also another option - induce a failure and re-run the pre-submit check.
Most of the time, we're not interested in must-gather logs if tests
pass. This option can save a lot of cloud resources (5-10 minutes for
every successful run).