Reset AlwaysRun to default before running AlwaysRunInjector
It might happen that "make jobs" re-uses some parts of configuration
from previous runs as it tries to be smart. However, when bumping
versions on our side where previously AlwaysRun was false (e.g. OCP 4.15
when 4.15 was still a candidate release), then "make jobs" will take the
value to the newly generated Prow jobs. This is not desired as we might
want to have 4.15 as the main test platform and run tests always.