Generate Konflux components for serverless operator
- Add support for adding components nudges automatically
- Add support for FBC builder pipeline to build operator index image

Push the SO components to the hack repository since when we cut a new
branch we could potentially have 2 components with the same "release
version" on 2 different branches and coordinating the merge of 2
pull requests on 2 different branches is challenging and difficult.

---

Tested with:
```
FIRST_KONFLUX_BRANCH=release-1.32 CONTAINER_ENGINE=docker make generate-ci-no-clean
```
while also enabling Konflux on Eventing 1.32 and 1.33 branches.

This produces:
- Eventing components with correct nudges for the
  `serverless-bundle-release-<xyz>`
- All components for serverless-operator, including the nudge for `serveless-bundle`
  to the `serverless-index`

Signed-off-by: Pierangelo Di Pilato <pierdipi@redhat.com>